Phase C6,C8: Move assembly/extern declarations out of signal.rs and thread.rs
- Move signal trampoline global_asm! and trampoline_phys_addr() to asm/mod.rs
- Move powersave extern declaration to asm::powersave_fn_addr()
- signal.rs re-exports trampoline_phys_addr from asm module

Modules losing #![allow(unsafe_code)]:
- processes/signal.rs
- processes/thread.rs
